+++ src/calva_backseat_driver/mcp/requests.cljs
@@ -1,8 +1,9 @@
 (ns calva-backseat-driver.mcp.requests
   (:require
    ["vscode" :as vscode]
-   [calva-backseat-driver.integrations.calva.features :as calva]
    [calva-backseat-driver.bracket-balance :as bracket-balance]
+   [calva-backseat-driver.integrations.calva.features :as calva]
+   [calva-backseat-driver.integrations.vscode.human-intelligence :as hi]
    [clojure.string :as string]
    [promesa.core :as p]))
 
@@ -144,6 +145,17 @@
                    :audience ["user" "assistant"]
                    :priority 10}}))
 
+(def human-intelligence-tool-listing
+  (let [tool-name "request_human_input"]
+    {:name tool-name
+     :description (tool-description tool-name)
+     :inputSchema {:type "object"
+                   :properties {"prompt" {:type "string"
+                                          :description (param-description tool-name "prompt")}}
+                   :required ["prompt"]
+                   :audience ["user" "assistant"]
+                   :priority 10}}))
+
 (defn handle-request-fn [{:ex/keys [dispatch!] :as options
                           :mcp/keys [repl-enabled?]}
                          {:keys [id method params] :as request}]
@@ -184,7 +196,10 @@
                                       (conj replace-top-level-form-tool-listing)
 
                                       true
-                                      (conj insert-top-level-form-tool-listing))}}]
+                                      (conj insert-top-level-form-tool-listing)
+
+                                      true
+                                      (conj human-intelligence-tool-listing))}}]
       response)
 
     (= method "resources/templates/list")
@@ -284,6 +299,15 @@
            :id id
            :result {:content [{:type "text"
                                :text (js/JSON.stringify (clj->js result))}]}})
+
+        (= tool "request_human_input")
+        (p/let [{:keys [prompt]} arguments
+                result (hi/request-human-input! (merge options
+                                                       {:cbd/prompt prompt}))]
+          {:jsonrpc "2.0"
+           :id id
+           :result {:content [{:type "text"
+                               :text result}]}})
 
         :else
         {:jsonrpc "2.0"